zoo/requirements.txt
python-dotenv
openai
bert-extractive-summarizer
transformers
torch
fastapi
uvicorn
pymongo[srv] 